+++ threshold_resource_estimator.hpp
@@ -2,6 +2,10 @@
 
 #include <mesos/module/resource_estimator.hpp>
 
+namespace os {
+    struct Load;
+}
+
 namespace com { namespace blue_yonder {
 
 class ThresholdResourceEstimatorProcess;
@@ -9,14 +13,24 @@ class ThresholdResourceEstimatorProcess;
 class ThresholdResourceEstimator : public mesos::slave::ResourceEstimator
 {
 public:
-    ThresholdResourceEstimator(mesos::Resources const & fixed);
+    ThresholdResourceEstimator(
+        std::function<Try<os::Load>()> const & load,
+        mesos::Resources const & fixed,
+        Option<double> const & loadThreshold1Min,
+        Option<double> const & loadThreshold5Min,
+        Option<double> const & loadThreshold15Min
+    );
     virtual Try<Nothing> initialize(const std::function<process::Future<mesos::ResourceUsage>()>&) final;
     virtual process::Future<mesos::Resources> oversubscribable() final;
     virtual ~ThresholdResourceEstimator();
 
 private:
     process::Owned<ThresholdResourceEstimatorProcess> process;
+    std::function<Try<os::Load>()> const load;
     mesos::Resources const fixed;
+    Option<double> const loadThreshold1Min;
+    Option<double> const loadThreshold5Min;
+    Option<double> const loadThreshold15Min;
 };
 
 } }